Support issuer filtering when matching rules (fixes #149) (#150)
* fix(loader): add support for issuer when matching rules

With Webpack v2 one can use svg-sprite-loader for JavaScript/JSX modules
while still using other loaders in other contexts (CSS for example).
This is possible by using the "issuer" configuration for loader rules.
For use with Webpack v1 an error will still occur.

Fixes #149

* test(loader): add coverage for Webpack v1+2 handling of issuer in rules

This adds coverage to the loader tests so that the "issuer" attribute on loader rules is honored
when evaluating Webpack v2 config. For v1 configuration the issuer attribute is ignored and the
previous errors are expected.